- Codebase organization and better maintainability. - Saving/loading now uses propery OS specific paths. - Saving/loading speedup and simplification. - Removed duplicate and cicular imports. - Simplified main.py so that it's only an entry point. - Moved the Model class into a dedicated module.
